lsp_server

Enum ErrorCode

#[non_exhaustive]
pub enum ErrorCode {

    ParseError = -32_700,
    InvalidRequest = -32_600,
    MethodNotFound = -32_601,
    InvalidParams = -32_602,
    InternalError = -32_603,
    ServerErrorStart = -32_099,
    ServerErrorEnd = -32_000,
    ServerNotInitialized = -32_002,
    UnknownErrorCode = -32_001,
    RequestCanceled = -32_800,
    ContentModified = -32_801,
    ServerCancelled = -32_802,
    RequestFailed = -32_803,

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ParseError = -32_700

InvalidRequest = -32_600

MethodNotFound = -32_601

InvalidParams = -32_602

InternalError = -32_603

ServerErrorStart = -32_099

ServerErrorEnd = -32_000

ServerNotInitialized = -32_002

Error code indicating that a server received a notification or request before the server has received the initialize request.

UnknownErrorCode = -32_001

RequestCanceled = -32_800

The client has canceled a request and a server has detected the cancel.

ContentModified = -32_801

The server detected that the content of a document got modified outside normal conditions. A server should NOT send this error code if it detects a content change in it unprocessed messages. The result even computed on an older state might still be useful for the client.

If a client decides that a result is not of any use anymore the client should cancel the request.

ServerCancelled = -32_802

The server cancelled the request. This error code should only be used for requests that explicitly support being server cancellable.

@since 3.17.0

RequestFailed = -32_803

A request failed but it was syntactically correct, e.g the method name was known and the parameters were valid. The error message should contain human readable information about why the request failed.

@since 3.17.0