Crate lsp_server

A language server scaffold, exposing a synchronous crossbeam-channel based API. This crate handles protocol handshaking and parsing messages, while you control the message dispatch loop yourself.

Run with RUST_LOG=lsp_server=debug to see all the messages.

Structs

Connection

Connection is just a pair of channels of LSP messages.

Incoming

IoThreads

Notification

Outgoing

ProtocolError

ReqQueue

Manages the set of pending requests, both incoming and outgoing.

Request

RequestId

Response

ResponseError

Enums

ErrorCode

ExtractError

Message